src/domains/notifications/domain/usecases/get-recipient-notifications-usecase/get-recipient-notifications-usecase.ts
import { BaseUsecase } from '@core/domain/usecases/base-usecase';
import { Injectable } from '@nestjs/common';
import { Notification } from '../../entities/notification/notification.entity';
import { NotificationsRepository } from '../../repositories/notifications-repository.interface';

interface GetRecipientNotificationsRequest {
  recipientId: string;
}

interface GetRecipientNotificationsResponse {
  notifications: Array<Notification>;
}

@Injectable()
export class GetRecipientNotificationsUsecase
  implements
    BaseUsecase<
      GetRecipientNotificationsRequest,
      GetRecipientNotificationsResponse
    >
{
  constructor(
    private readonly notificationsRepository: NotificationsRepository,
  ) {}

  async execute(
    request: GetRecipientNotificationsRequest,
  ): Promise<GetRecipientNotificationsResponse> {
    const { recipientId } = request;

    const notifications =
      await this.notificationsRepository.findManyByRecipientId(recipientId);

    return { notifications };
  }
}
